requirements.txt
streamlit
pandas
numpy
python-dotenv
s3fs
boto3

